pyproject.toml
[tool.poetry]
name = "opentpod-tools"
version = "0.1.0"
description = "Command line tools integrating OpenTPOD functionality."
authors = [
    "Jan Harkes <jaharkes@cs.cmu.edu>",
]
license = "Apache-2.0"
repository = "https://github.com/cmusatyalab/opentpod-tools/"

[tool.poetry.dependencies]
python = "^3.7.1"
ConfigArgParse = "^1.2.3"
datumaro = [
    {git = "https://github.com/openvinotoolkit/datumaro", rev = "v0.1.3", extras = ["tf"], markers = "extra != 'tf-gpu'"},
    {git = "https://github.com/openvinotoolkit/datumaro", rev = "v0.1.3", extras = ["tf-gpu"], markers = "extra == 'tf-gpu'"},
]
requests = "^2.24.0"
tqdm = "^4.50.2"
imagehash = "^4.1.0"
torch = "^1.7.0"
torchvision = "^0.8.1"
tensorflow = {version = "~1.15.4", optional = true }
tensorflow-gpu = {version = "~1.15.4", optional = true }

[tool.poetry.extras]
tf = ["tensorflow"]
tf-gpu = ["tensorflow-gpu"]

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pylint = "^2.6.0"
pytest = "^5.2"
black = "^20.8b1"
isort = "^5.6.4"
google-cloud-storage = "^1.33.0"
google-cloud-automl = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tpod-download = "opentpod_tools.download:main"
tpod-merge = "opentpod_tools.merge:main"
tpod-unique = "opentpod_tools.removedup:main"
tpod-class = "opentpod_tools.classification:main"
tpod-pytorch-class = "opentpod_tools.pytorch_classification:main"
tpod-pytorch-class-test = "opentpod_tools.pytorch_class_result:main"
tpod-google-automl-od = "opentpod_tools.google_automl_od:main"
tpod-google-automl-od-train = "opentpod_tools.google_automl_od_train:main"
tpod-google-automl-class = "opentpod_tools.google_automl_class:main"
tpod-google-automl-class-train = "opentpod_tools.google_automl_class_train:main"

[tool.poetry.urls]
issues = "https://github.com/cmusatyalab/opentpod-tools/issues"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88